requirements.txt
flask>=3.0,<4.0
pytest>=8.2,<9.0
pytest-html>=4.1,<5.0
pytest-metadata>=3.1,<4.0
playwright>=1.45,<2.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
coverage>=7.5,<8.0
